Add files via upload
Version 22 has been running for a week now without problems.
Circuit was modified by inserting a diode between the power supply 5V and the Arduino 5V to prevent current coming from the Arduino going into the power supply.  This happens when the Arduino is powered by the USB connection while programming.  In this situation, without this diode, the power supply will become non-functional.

Signed-off-by: OldSurferDude <github@greggferry.com>